+++ app/src/main/java/online/cyberforensic/wirelessscannerandroid/model/Dot11Device.java
@@ -1,10 +1,12 @@
 
 package online.cyberforensic.wirelessscannerandroid.model;
 
-import java.util.List;
-import javax.annotation.Generated;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
+
+import java.util.List;
+
+import javax.annotation.Generated;
 
 @Generated("jsonschema2pojo")
 public class Dot11Device {
@@ -17,82 +19,82 @@ public class Dot11Device {
     private String dot11DeviceLastBssid;
     @SerializedName("dot11.device.response_fingerprint")
     @Expose
-    private Integer dot11DeviceResponseFingerprint;
+    private Long dot11DeviceResponseFingerprint;
     @SerializedName("dot11.device.probe_fingerprint")
     @Expose
-    private Integer dot11DeviceProbeFingerprint;
+    private Long dot11DeviceProbeFingerprint;
     @SerializedName("dot11.device.beacon_fingerprint")
     @Expose
-    private Integer dot11DeviceBeaconFingerprint;
+    private Long dot11DeviceBeaconFingerprint;
     @SerializedName("dot11.device.neighbor_report_capable")
     @Expose
-    private Integer dot11DeviceNeighborReportCapable;
+    private Long dot11DeviceNeighborReportCapable;
     @SerializedName("dot11.device.link_measurement_capable")
     @Expose
-    private Integer dot11DeviceLinkMeasurementCapable;
+    private Long dot11DeviceLinkMeasurementCapable;
     @SerializedName("dot11.device.wps_m3_last")
     @Expose
-    private Integer dot11DeviceWpsM3Last;
+    private Long dot11DeviceWpsM3Last;
     @SerializedName("dot11.device.max_tx_power")
     @Expose
-    private Integer dot11DeviceMaxTxPower;
+    private Long dot11DeviceMaxTxPower;
     @SerializedName("dot11.device.wps_m3_count")
     @Expose
-    private Integer dot11DeviceWpsM3Count;
+    private Long dot11DeviceWpsM3Count;
     @SerializedName("dot11.device.min_tx_power")
     @Expose
-    private Integer dot11DeviceMinTxPower;
+    private Long dot11DeviceMinTxPower;
     @SerializedName("dot11.device.last_beacon_timestamp")
     @Expose
-    private Integer dot11DeviceLastBeaconTimestamp;
+    private Long dot11DeviceLastBeaconTimestamp;
     @SerializedName("dot11.device.datasize_retry")
     @Expose
-    private Integer dot11DeviceDatasizeRetry;
+    private Long dot11DeviceDatasizeRetry;
     @SerializedName("dot11.device.probed_ssid_map")
     @Expose
     private List<Dot11DeviceProbedSsidMap> dot11DeviceProbedSsidMap = null;
     @SerializedName("dot11.device.last_sequence")
     @Expose
-    private Integer dot11DeviceLastSequence;
+    private Long dot11DeviceLastSequence;
     @SerializedName("dot11.device.num_client_aps")
     @Expose
-    private Integer dot11DeviceNumClientAps;
+    private Long dot11DeviceNumClientAps;
     @SerializedName("dot11.device.num_advertised_ssids")
     @Expose
-    private Integer dot11DeviceNumAdvertisedSsids;
+    private Long dot11DeviceNumAdvertisedSsids;
     @SerializedName("dot11.device.num_probed_ssids")
     @Expose
-    private Integer dot11DeviceNumProbedSsids;
+    private Long dot11DeviceNumProbedSsids;
     @SerializedName("dot11.device.num_associated_clients")
     @Expose
-    private Integer dot11DeviceNumAssociatedClients;
+    private Long dot11DeviceNumAssociatedClients;
     @SerializedName("dot11.device.wpa_present_handshake")
     @Expose
-    private Integer dot11DeviceWpaPresentHandshake;
+    private Long dot11DeviceWpaPresentHandshake;
     @SerializedName("dot11.device.client_disconnects")
     @Expose
-    private Integer dot11DeviceClientDisconnects;
+    private Long dot11DeviceClientDisconnects;
     @SerializedName("dot11.device.typeset")
     @Expose
-    private Integer dot11DeviceTypeset;
+    private Long dot11DeviceTypeset;
     @SerializedName("dot11.device.num_responded_ssids")
     @Expose
-    private Integer dot11DeviceNumRespondedSsids;
+    private Long dot11DeviceNumRespondedSsids;
     @SerializedName("dot11.device.client_disconnects_last")
     @Expose
-    private Integer dot11DeviceClientDisconnectsLast;
+    private Long dot11DeviceClientDisconnectsLast;
     @SerializedName("dot11.device.bss_timestamp")
     @Expose
-    private Integer dot11DeviceBssTimestamp;
+    private Long dot11DeviceBssTimestamp;
     @SerializedName("dot11.device.num_fragments")
     @Expose
-    private Integer dot11DeviceNumFragments;
+    private Long dot11DeviceNumFragments;
     @SerializedName("dot11.device.num_retries")
     @Expose
-    private Integer dot11DeviceNumRetries;
+    private Long dot11DeviceNumRetries;
     @SerializedName("dot11.device.datasize")
     @Expose
-    private Integer dot11DeviceDatasize;
+    private Long dot11DeviceDatasize;
     @SerializedName("dot11.device.ssid_beacon_packet")
     @Expose
     private Dot11DeviceSsidBeaconPacket dot11DeviceSsidBeaconPacket;
@@ -131,91 +133,91 @@ public class Dot11Device {
         this.dot11DeviceLastBssid = dot11DeviceLastBssid;
     }
 
-    public Integer getDot11DeviceResponseFingerprint() {
+    public Long getDot11DeviceResponseFingerprint() {
         return dot11DeviceResponseFingerprint;
     }
 
-    public void setDot11DeviceResponseFingerprint(Integer dot11DeviceResponseFingerprint) {
+    public void setDot11DeviceResponseFingerprint(Long dot11DeviceResponseFingerprint) {
         this.dot11DeviceResponseFingerprint = dot11DeviceResponseFingerprint;
     }
 
-    public Integer getDot11DeviceProbeFingerprint() {
+    public Long getDot11DeviceProbeFingerprint() {
         return dot11DeviceProbeFingerprint;
     }
 
-    public void setDot11DeviceProbeFingerprint(Integer dot11DeviceProbeFingerprint) {
+    public void setDot11DeviceProbeFingerprint(Long dot11DeviceProbeFingerprint) {
         this.dot11DeviceProbeFingerprint = dot11DeviceProbeFingerprint;
     }
 
-    public Integer getDot11DeviceBeaconFingerprint() {
+    public Long getDot11DeviceBeaconFingerprint() {
         return dot11DeviceBeaconFingerprint;
     }
 
-    public void setDot11DeviceBeaconFingerprint(Integer dot11DeviceBeaconFingerprint) {
+    public void setDot11DeviceBeaconFingerprint(Long dot11DeviceBeaconFingerprint) {
         this.dot11DeviceBeaconFingerprint = dot11DeviceBeaconFingerprint;
     }
 
-    public Integer getDot11DeviceNeighborReportCapable() {
+    public Long getDot11DeviceNeighborReportCapable() {
         return dot11DeviceNeighborReportCapable;
     }
 
-    public void setDot11DeviceNeighborReportCapable(Integer dot11DeviceNeighborReportCapable) {
+    public void setDot11DeviceNeighborReportCapable(Long dot11DeviceNeighborReportCapable) {
         this.dot11DeviceNeighborReportCapable = dot11DeviceNeighborReportCapable;
     }
 
-    public Integer getDot11DeviceLinkMeasurementCapable() {
+    public Long getDot11DeviceLinkMeasurementCapable() {
         return dot11DeviceLinkMeasurementCapable;
     }
 
-    public void setDot11DeviceLinkMeasurementCapable(Integer dot11DeviceLinkMeasurementCapable) {
+    public void setDot11DeviceLinkMeasurementCapable(Long dot11DeviceLinkMeasurementCapable) {
         this.dot11DeviceLinkMeasurementCapable = dot11DeviceLinkMeasurementCapable;
     }
 
-    public Integer getDot11DeviceWpsM3Last() {
+    public Long getDot11DeviceWpsM3Last() {
         return dot11DeviceWpsM3Last;
     }
 
-    public void setDot11DeviceWpsM3Last(Integer dot11DeviceWpsM3Last) {
+    public void setDot11DeviceWpsM3Last(Long dot11DeviceWpsM3Last) {
         this.dot11DeviceWpsM3Last = dot11DeviceWpsM3Last;
     }
 
-    public Integer getDot11DeviceMaxTxPower() {
+    public Long getDot11DeviceMaxTxPower() {
         return dot11DeviceMaxTxPower;
     }
 
-    public void setDot11DeviceMaxTxPower(Integer dot11DeviceMaxTxPower) {
+    public void setDot11DeviceMaxTxPower(Long dot11DeviceMaxTxPower) {
         this.dot11DeviceMaxTxPower = dot11DeviceMaxTxPower;
     }
 
-    public Integer getDot11DeviceWpsM3Count() {
+    public Long getDot11DeviceWpsM3Count() {
         return dot11DeviceWpsM3Count;
     }
 
-    public void setDot11DeviceWpsM3Count(Integer dot11DeviceWpsM3Count) {
+    public void setDot11DeviceWpsM3Count(Long dot11DeviceWpsM3Count) {
         this.dot11DeviceWpsM3Count = dot11DeviceWpsM3Count;
     }
 
-    public Integer getDot11DeviceMinTxPower() {
+    public Long getDot11DeviceMinTxPower() {
         return dot11DeviceMinTxPower;
     }
 
-    public void setDot11DeviceMinTxPower(Integer dot11DeviceMinTxPower) {
+    public void setDot11DeviceMinTxPower(Long dot11DeviceMinTxPower) {
         this.dot11DeviceMinTxPower = dot11DeviceMinTxPower;
     }
 
-    public Integer getDot11DeviceLastBeaconTimestamp() {
+    public Long getDot11DeviceLastBeaconTimestamp() {
         return dot11DeviceLastBeaconTimestamp;
     }
 
-    public void setDot11DeviceLastBeaconTimestamp(Integer dot11DeviceLastBeaconTimestamp) {
+    public void setDot11DeviceLastBeaconTimestamp(Long dot11DeviceLastBeaconTimestamp) {
         this.dot11DeviceLastBeaconTimestamp = dot11DeviceLastBeaconTimestamp;
     }
 
-    public Integer getDot11DeviceDatasizeRetry() {
+    public Long getDot11DeviceDatasizeRetry() {
         return dot11DeviceDatasizeRetry;
     }
 
-    public void setDot11DeviceDatasizeRetry(Integer dot11DeviceDatasizeRetry) {
+    public void setDot11DeviceDatasizeRetry(Long dot11DeviceDatasizeRetry) {
         this.dot11DeviceDatasizeRetry = dot11DeviceDatasizeRetry;
     }
 
@@ -227,115 +229,115 @@ public class Dot11Device {
         this.dot11DeviceProbedSsidMap = dot11DeviceProbedSsidMap;
     }
 
-    public Integer getDot11DeviceLastSequence() {
+    public Long getDot11DeviceLastSequence() {
         return dot11DeviceLastSequence;
     }
 
-    public void setDot11DeviceLastSequence(Integer dot11DeviceLastSequence) {
+    public void setDot11DeviceLastSequence(Long dot11DeviceLastSequence) {
         this.dot11DeviceLastSequence = dot11DeviceLastSequence;
     }
 
-    public Integer getDot11DeviceNumClientAps() {
+    public Long getDot11DeviceNumClientAps() {
         return dot11DeviceNumClientAps;
     }
 
-    public void setDot11DeviceNumClientAps(Integer dot11DeviceNumClientAps) {
+    public void setDot11DeviceNumClientAps(Long dot11DeviceNumClientAps) {
         this.dot11DeviceNumClientAps = dot11DeviceNumClientAps;
     }
 
-    public Integer getDot11DeviceNumAdvertisedSsids() {
+    public Long getDot11DeviceNumAdvertisedSsids() {
         return dot11DeviceNumAdvertisedSsids;
     }
 
-    public void setDot11DeviceNumAdvertisedSsids(Integer dot11DeviceNumAdvertisedSsids) {
+    public void setDot11DeviceNumAdvertisedSsids(Long dot11DeviceNumAdvertisedSsids) {
         this.dot11DeviceNumAdvertisedSsids = dot11DeviceNumAdvertisedSsids;
     }
 
-    public Integer getDot11DeviceNumProbedSsids() {
+    public Long getDot11DeviceNumProbedSsids() {
         return dot11DeviceNumProbedSsids;
     }
 
-    public void setDot11DeviceNumProbedSsids(Integer dot11DeviceNumProbedSsids) {
+    public void setDot11DeviceNumProbedSsids(Long dot11DeviceNumProbedSsids) {
         this.dot11DeviceNumProbedSsids = dot11DeviceNumProbedSsids;
     }
 
-    public Integer getDot11DeviceNumAssociatedClients() {
+    public Long getDot11DeviceNumAssociatedClients() {
         return dot11DeviceNumAssociatedClients;
     }
 
-    public void setDot11DeviceNumAssociatedClients(Integer dot11DeviceNumAssociatedClients) {
+    public void setDot11DeviceNumAssociatedClients(Long dot11DeviceNumAssociatedClients) {
         this.dot11DeviceNumAssociatedClients = dot11DeviceNumAssociatedClients;
     }
 
-    public Integer getDot11DeviceWpaPresentHandshake() {
+    public Long getDot11DeviceWpaPresentHandshake() {
         return dot11DeviceWpaPresentHandshake;
     }
 
-    public void setDot11DeviceWpaPresentHandshake(Integer dot11DeviceWpaPresentHandshake) {
+    public void setDot11DeviceWpaPresentHandshake(Long dot11DeviceWpaPresentHandshake) {
         this.dot11DeviceWpaPresentHandshake = dot11DeviceWpaPresentHandshake;
     }
 
-    public Integer getDot11DeviceClientDisconnects() {
+    public Long getDot11DeviceClientDisconnects() {
         return dot11DeviceClientDisconnects;
     }
 
-    public void setDot11DeviceClientDisconnects(Integer dot11DeviceClientDisconnects) {
+    public void setDot11DeviceClientDisconnects(Long dot11DeviceClientDisconnects) {
         this.dot11DeviceClientDisconnects = dot11DeviceClientDisconnects;
     }
 
-    public Integer getDot11DeviceTypeset() {
+    public Long getDot11DeviceTypeset() {
         return dot11DeviceTypeset;
     }
 
-    public void setDot11DeviceTypeset(Integer dot11DeviceTypeset) {
+    public void setDot11DeviceTypeset(Long dot11DeviceTypeset) {
         this.dot11DeviceTypeset = dot11DeviceTypeset;
     }
 
-    public Integer getDot11DeviceNumRespondedSsids() {
+    public Long getDot11DeviceNumRespondedSsids() {
         return dot11DeviceNumRespondedSsids;
     }
 
-    public void setDot11DeviceNumRespondedSsids(Integer dot11DeviceNumRespondedSsids) {
+    public void setDot11DeviceNumRespondedSsids(Long dot11DeviceNumRespondedSsids) {
         this.dot11DeviceNumRespondedSsids = dot11DeviceNumRespondedSsids;
     }
 
-    public Integer getDot11DeviceClientDisconnectsLast() {
+    public Long getDot11DeviceClientDisconnectsLast() {
         return dot11DeviceClientDisconnectsLast;
     }
 
-    public void setDot11DeviceClientDisconnectsLast(Integer dot11DeviceClientDisconnectsLast) {
+    public void setDot11DeviceClientDisconnectsLast(Long dot11DeviceClientDisconnectsLast) {
         this.dot11DeviceClientDisconnectsLast = dot11DeviceClientDisconnectsLast;
     }
 
-    public Integer getDot11DeviceBssTimestamp() {
+    public Long getDot11DeviceBssTimestamp() {
         return dot11DeviceBssTimestamp;
     }
 
-    public void setDot11DeviceBssTimestamp(Integer dot11DeviceBssTimestamp) {
+    public void setDot11DeviceBssTimestamp(Long dot11DeviceBssTimestamp) {
         this.dot11DeviceBssTimestamp = dot11DeviceBssTimestamp;
     }
 
-    public Integer getDot11DeviceNumFragments() {
+    public Long getDot11DeviceNumFragments() {
         return dot11DeviceNumFragments;
     }
 
-    public void setDot11DeviceNumFragments(Integer dot11DeviceNumFragments) {
+    public void setDot11DeviceNumFragments(Long dot11DeviceNumFragments) {
         this.dot11DeviceNumFragments = dot11DeviceNumFragments;
     }
 
-    public Integer getDot11DeviceNumRetries() {
+    public Long getDot11DeviceNumRetries() {
         return dot11DeviceNumRetries;
     }
 
-    public void setDot11DeviceNumRetries(Integer dot11DeviceNumRetries) {
+    public void setDot11DeviceNumRetries(Long dot11DeviceNumRetries) {
         this.dot11DeviceNumRetries = dot11DeviceNumRetries;
     }
 
-    public Integer getDot11DeviceDatasize() {
+    public Long getDot11DeviceDatasize() {
         return dot11DeviceDatasize;
     }
 
-    public void setDot11DeviceDatasize(Integer dot11DeviceDatasize) {
+    public void setDot11DeviceDatasize(Long dot11DeviceDatasize) {
         this.dot11DeviceDatasize = dot11DeviceDatasize;
     }
